+++ swarm/src/behaviour/toggle.rs
@@ -31,7 +31,8 @@ use crate::{
     THandlerOutEvent,
 };
 use either::Either;
-use libp2p_core::{either::EitherOutput, upgrade::DeniedUpgrade, Endpoint, Multiaddr, PeerId};
+use futures::future;
+use libp2p_core::{upgrade::DeniedUpgrade, Endpoint, Multiaddr, PeerId};
 use std::error::Error;
 use std::{task::Context, task::Poll};
 
@@ -210,8 +211,8 @@ where
         >,
     ) {
         let out = match out {
-            EitherOutput::First(out) => out,
-            EitherOutput::Second(v) => void::unreachable(v),
+            future::Either::Left(out) => out,
+            future::Either::Right(v) => void::unreachable(v),
         };
 
         if let Either::Left(info) = info {